9cc.cファイルを削除し、解析ロジックを新たにparse.cに実装しました。Makefileを修正し、静的リンクオプションをLDFLAGSに移動しました。
