+++ 9cc/Makefile
@@ -1,4 +1,5 @@
-CFLAGS=-std=c11 -g -static
+CFLAGS=-std=c11 -g
+LDFLAGS=-static
 SRCS=$(wildcard *.c)
 OBJS=$(SRCS:.c=.o)
